Move validators into their own package and modules
