+++ spectacles/validators/sql.py
@@ -1,20 +1,12 @@
+from typing import Union, Dict, Any, List, Optional
 import time
-from typing import Any, List, Dict, Union, Optional
-from abc import ABC, abstractmethod
-from collections import OrderedDict
+from spectacles.validators.validator import Validator
 from spectacles.client import LookerClient
-from spectacles.lookml import Project, Model, Explore, Dimension
+from spectacles.lookml import Project, Model, Dimension, Explore
 from spectacles.types import QueryMode
+from spectacles.exceptions import SpectaclesException, SqlError, LookMlNotFound
 from spectacles.select import is_selected
 from spectacles.logger import GLOBAL_LOGGER as logger
-from spectacles.exceptions import (
-    SqlError,
-    DataTestError,
-    ContentError,
-    SpectaclesException,
-    LookMlNotFound,
-)
-import spectacles.printer as printer
 
 
 class Query:
@@ -42,262 +34,6 @@ class QueryResult:
         self.query_task_id = query_task_id
         self.status = status
         self.error = error
-
-
-class Validator(ABC):  # pragma: no cover
-    """Defines abstract base interface for validators.
-
-    Not intended to be used directly, only inherited.
-
-    Attributes:
-        client: Looker API client.
-
-    """
-
-    def __init__(self, client: LookerClient):
-        self.client = client
-
-    @abstractmethod
-    def validate(self):
-        raise NotImplementedError
-
-
-class ContentValidator(Validator):
-    def __init__(
-        self, client: LookerClient, project: str, exclude_personal: bool = False
-    ):
-        super().__init__(client)
-        self.project = Project(project, models=[])
-        personal_folders = self._get_personal_folders() if exclude_personal else []
-        self.personal_folders: List[int] = personal_folders
-
-    def _get_personal_folders(self) -> List[int]:
-        personal_folders = []
-        result = self.client.all_folders(self.project.name)
-        for folder in result:
-            if folder["is_personal"] or folder["is_personal_descendant"]:
-                personal_folders.append(folder["id"])
-        return personal_folders
-
-    def build_project(
-        self,
-        selectors: Optional[List[str]] = None,
-        exclusions: Optional[List[str]] = None,
-    ) -> None:
-        """Creates an object representation of the project's LookML.
-
-        Args:
-            selectors: List of selector strings in 'model_name/explore_name' format.
-                The '*' wildcard selects all models or explores. For instance,
-                'model_name/*' would select all explores in the 'model_name' model.
-
-        """
-        # Assign default values for selectors and exclusions
-        if selectors is None:
-            selectors = ["*/*"]
-        if exclusions is None:
-            exclusions = []
-
-        all_models = [
-            Model.from_json(model) for model in self.client.get_lookml_models()
-        ]
-        project_models = [
-            model for model in all_models if model.project_name == self.project.name
-        ]
-
-        if not project_models:
-            raise LookMlNotFound(
-                name="project-models-not-found",
-                title="No configured models found for the specified project.",
-                detail=(
-                    f"Go to {self.client.base_url}/projects and confirm "
-                    "a) at least one model exists for the project and "
-                    "b) it has an active configuration."
-                ),
-            )
-
-        for model in project_models:
-            model.explores = [
-                explore
-                for explore in model.explores
-                if is_selected(model.name, explore.name, selectors, exclusions)
-            ]
-
-        self.project.models = [
-            model for model in project_models if len(model.explores) > 0
-        ]
-
-    def validate(self) -> Dict[str, Any]:
-        result = self.client.content_validation()
-        self.project.queried = True
-
-        for content in result["content_with_errors"]:
-            # Skip content dicts if they lack a `look` or `dashboard` key
-            try:
-                content_type = self._get_content_type(content)
-            except KeyError:
-                logger.debug(
-                    f"Skipping content because it does not seem to be a dashboard or "
-                    f"a look. The content received was: {content}"
-                )
-                continue
-
-            # If exclude_personal isn't specified, personal_folders list is empty
-            if content[content_type]["folder"]["id"] in self.personal_folders:
-                continue
-            else:
-                self._handle_content_result(content, content_type)
-
-        return self.project.get_results(validator="content")
-
-    @staticmethod
-    def _get_content_type(content: Dict[str, Any]) -> str:
-        if content["dashboard"]:
-            return "dashboard"
-        elif content["look"]:
-            return "look"
-        else:
-            raise KeyError("Content type not found. Valid keys are 'look', 'dashboard'")
-
-    def _handle_content_result(self, content: Dict, content_type: str) -> None:
-        for error in content["errors"]:
-            model_name = error["model_name"]
-            explore_name = error["explore_name"]
-            explore = self.project.get_explore(model=model_name, name=explore_name)
-            # Skip errors that are not associated with selected explores
-            if explore is None:
-                continue
-
-            content_id = content[content_type]["id"]
-            content_error = ContentError(
-                model=model_name,
-                explore=explore_name,
-                message=error["message"],
-                field_name=error["field_name"],
-                content_type=content_type,
-                title=content[content_type]["title"],
-                space=content[content_type]["space"]["name"],
-                url=f"{self.client.base_url}/{content_type}s/{content_id}",
-            )
-            if content_error not in explore.errors:
-                explore.errors.append(content_error)
-
-
-class DataTestValidator(Validator):
-    """Runs LookML/data tests for a given project.
-
-    Args:
-        client: Looker API client.
-        project: Name of the LookML project to validate.
-
-    """
-
-    def __init__(self, client: LookerClient, project: str):
-        super().__init__(client)
-        self.project = project
-
-    def validate(
-        self,
-        selectors: Optional[List[str]] = None,
-        exclusions: Optional[List[str]] = None,
-    ) -> Dict[str, Any]:
-        # Assign default values for selectors and exclusions
-        if selectors is None:
-            selectors = ["*/*"]
-        if exclusions is None:
-            exclusions = []
-
-        all_tests = self.client.all_lookml_tests(self.project)
-        selected_tests = []
-        test_to_explore = {}
-        for test in all_tests:
-            if is_selected(
-                test["model_name"], test["explore_name"], selectors, exclusions
-            ):
-                selected_tests.append(test)
-                # The error objects don't contain the name of the explore
-                # We create this mapping to help look up the explore from the test name
-                test_to_explore[test["name"]] = test["explore_name"]
-
-        test_count = len(selected_tests)
-        if test_count == 0:
-            raise SpectaclesException(
-                name="no-data-tests-found",
-                title="No data tests found.",
-                detail=(
-                    "If you're using --explores or --exclude, make sure your project "
-                    "has data tests that reference those models or explores."
-                ),
-            )
-
-        printer.print_header(
-            f"Running {test_count} {'test' if test_count == 1 else 'tests'}"
-        )
-
-        test_results: List[Dict[str, Any]] = []
-        for test in selected_tests:
-            test_name = test["name"]
-            model_name = test["model_name"]
-            results = self.client.run_lookml_test(
-                self.project, model=model_name, test=test_name
-            )
-            test_results.extend(results)
-
-        tested = []
-        errors = []
-
-        for result in test_results:
-            explore = test_to_explore[result["test_name"]]
-            test = {
-                "model": result["model_name"],
-                "explore": explore,
-                "passed": result["success"],
-            }
-            tested.append(test)
-            if not result["success"]:
-                for error in result["errors"]:
-                    project, file_path = error["file_path"].split("/", 1)
-                    lookml_url = (
-                        f"{self.client.base_url}/projects/{self.project}"
-                        f"/files/{file_path}?line={error['line_number']}"
-                    )
-                    errors.append(
-                        DataTestError(
-                            model=error["model_id"],
-                            explore=error["explore"],
-                            message=error["message"],
-                            test_name=result["test_name"],
-                            lookml_url=lookml_url,
-                        ).__dict__
-                    )
-
-        def reduce_result(results):
-            """Aggregate individual test results to get pass/fail by explore"""
-            agg = OrderedDict()
-            for result in results:
-                # Keys by model and explore, adds additional values for `passed` to a set
-                agg.setdefault((result["model"], result["explore"]), set()).add(
-                    result["passed"]
-                )
-            reduced = [
-                {"model": k[0], "explore": k[1], "passed": min(v)}
-                for k, v in agg.items()
-            ]
-            return reduced
-
-        tested = reduce_result(tested)
-        for test in tested:
-            printer.print_validation_result(
-                passed=test["passed"], source=f"{test['model']}.{test['explore']}"
-            )
-
-        passed = min((test["passed"] for test in tested), default=True)
-        return {
-            "validator": "data_test",
-            "status": "passed" if passed else "failed",
-            "tested": tested,
-            "errors": errors,
-        }
 
 
 class SqlValidator(Validator):
